contest: move validations to utils dir
